# docs/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "NodeHub -  API 文档\n\n这是 NodeHub 的 API 文档\n\n## 认证\n大多数接口需要使用 JWT 令牌进行认证。\n认证时，请在 Authorization 头中包含 JWT 令牌：\n`Authorization: Bearer <your-jwt-token>`\n\n## 错误响应\n所有错误响应都遵循统一格式，包含 code、message 和 error 字段。\n\n## 成功响应\n所有成功响应都遵循统一格式，包含 code、message 和 data 字段。",
    "title": "NodeHub API",
    "contact": {
      "name": "NodeHub API 支持",
      "email": "support@nodehub.com"
    },
    "license": {
      "name": "GPL-3.0",
      "url": "https://opensource.org/license/gpl-3-0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/update": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "获取程序最新版本信息",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "更新"
        ],
        "summary": "最新版本",
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "additionalProperties": {
                        "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_core_update.LatestInfo"
                      }
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
    "/api/v1/update/:name": {
      "post": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "description": "更新程序",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "更新"
        ],
        "summary": "更新",
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
    "/api/v1/sub/log": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "订阅"
        ],
        "summary": "获取订阅运行日志",
        "description": "获取订阅拉取/转换日志，包含运行记录与事件",
        "parameters": [
          {
            "type": "integer",
            "description": "订阅ID",
            "name": "sub_id",
            "in": "query",
            "required": true
          },
          {
            "type": "integer",
            "description": "返回条数",
            "name": "limit",
            "in": "query"
          },
          {
            "name": "include_events",
            "in": "query",
            "description": "是否包含历史 events",
            "required": false,
            "type": "boolean",
            "default": false
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub_RunLogResponse"
                    }
                  }
                }
              ]
            }
          },
          "400": {
            "description": "请求参数错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "401": {
            "description": "未授权",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          },
          "500": {
            "description": "服务器内部错误",
            "schema": {
              "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "github_com_evilCYH_nodehub_internal_core_update.LatestInfo": {
      "type": "object",
      "properties": {
        "body": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "published_at": {
          "type": "string"
        },
        "tag_name": {
          "type": "string"
        }
      }
    },
...
    "github_com_evilCYH_nodehub_internal_server_resp.ResponseStruct": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "example": 200
        },
        "data": {},
        "message": {
          "type": "string",
          "example": "success"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub_RunLog": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "sub_id": {
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "raw_count": {
          "type": "integer"
        },
        "accepted": {
          "type": "integer"
        },
        "created_at": {
          "type": "string"
        },
        "duration_ms": {
          "type": "integer"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub_RunEvent": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "run_id": {
          "type": "integer",
          "format": "int64"
        },
        "sub_id": {
          "type": "integer"
        },
        "step": {
          "type": "string"
        },
        "level": {
          "type": "string"
        },
        "message": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "run_time": {
          "type": "string"
        }
      }
    },
    "github_com_evilCYH_nodehub_internal_models_sub_RunLogResponse": {
      "type": "object",
      "properties": {
        "runs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub_RunLog"
          }
        },
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/github_com_evilCYH_nodehub_internal_models_sub_RunEvent"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "类型为 \"Bearer\"，后跟空格和 JWT 令牌。",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}